# build/swagger.json
{
	"components": {
...
		"schemas": {
...
			"HandleUserInputRequest": {
				"properties": {
					"userInput": {
						"type": "string"
					},
					"assistantId": {
						"type": "string"
					},
					"threadId": {
						"type": "string"
					}
				},
				"required": [
					"userInput",
					"assistantId",
					"threadId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GenerateAudioRequest": {
				"properties": {
					"text": {
						"type": "string"
					}
				},
				"required": [
					"text"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "my-task-manager",
		"version": "1.0.0",
		"description": "API service for managing tasks",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/assistant/user-input": {
			"post": {
				"operationId": "HandleUserInput",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/HandleUserInputRequest"
							}
						}
					}
				}
			}
		},
		"/audio/generate": {
			"post": {
				"operationId": "GenerateAudio",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenerateAudioRequest"
							}
						}
					}
				}
			}
		},
		"/audio/{id}": {
			"get": {
				"operationId": "GetAudio",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "https://sb-api.ngrok.app/"
		}
	]
}